+++ components/action_buttons/action_buttons.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import { actionPlayerHit, actionPlayerStand } from "../scripts/game.js";
+import { actionPlayerHit, actionPlayerStand } from "../../scripts/game.js";
 
 
 export function setupActionButtons() {
