+++ components/card/card.js
@@ -1,11 +1,11 @@
 // @ts-check
 
-import { constants } from "../index.js";
-import { getPileLocation } from "./positions.js";
+import { constants } from "../../index.js";
+import { getPileLocation } from "../positions.js";
 
 
 /**
- * @import {Card} from "../scripts/cards"
+ * @import {Card} from "../../scripts/cards"
  */
