+++ components/deck/deck.js
@@ -1,7 +1,7 @@
 // @ts-check
 
-import { constants } from "../index.js";
-import { getPileLocation } from "./positions.js";
+import { constants } from "../../index.js";
+import { getPileLocation } from "../positions.js";
 
 
 /**
