+++ index.js
@@ -1,11 +1,11 @@
 // @ts-check
 
 import { Constants } from "./components/constants.js";
-import { renderPhantomDeck } from "./components/deck.js";
+import { renderPhantomDeck } from "./components/deck/deck.js";
 import { getPlayerBank, resetGameState } from "./scripts/state.js";
-import { setupModal } from "./components/modal.js";
-import { setupActionButtons } from "./components/action_buttons.js";
-import { renderPlayerBank } from "./components/bank.js";
+import { setupModal } from "./components/modal/modal.js";
+import { setupActionButtons } from "./components/action_buttons/action_buttons.js";
+import { renderPlayerBank } from "./components/bank/bank.js";
 
 /**
  * @type {Constants}
